+++ dev-setup.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+python3 -m venv .venv
+source .venv/bin/activate
+pip install -e ".[dev,test,docs]"
+pre-commit install