+++ dev-test.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+flake8 pip_chill tests
+pytest
+tox