+++ docs/conf.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 #
 # pip_chill documentation build configuration file, created by
 # sphinx-quickstart on Tue Jul  9 22:26:36 2013.
@@ -13,8 +12,8 @@
 # All configuration values have a default; values that are commented out
 # serve to show the default.
 
-import sys
 import os
+import sys
 
 # If extensions (or modules to document with autodoc) are in another
 # directory, add these directories to sys.path here. If the directory is
@@ -31,7 +30,7 @@ project_root = os.path.dirname(cwd)
 # version is used.
 sys.path.insert(0, project_root)
 
-import pip_chill
+import pip_chill  # noqa: E402
 
 # -- General configuration ---------------------------------------------
 
@@ -55,8 +54,8 @@ source_suffix = ".rst"
 master_doc = "index"
 
 # General information about the project.
-project = u"PIP Chill"
-copyright = u"2016, Ricardo Bánffy"
+project = "PIP Chill"
+COPYRIGHT = "2016, Ricardo Bánffy"
 
 # The version info for the project you're documenting, acts as replacement
 # for |version| and |release|, also used in various other places throughout
@@ -195,11 +194,11 @@ htmlhelp_basename = "pip_chilldoc"
 
 latex_elements = {
     # The paper size ('letterpaper' or 'a4paper').
-    #'papersize': 'letterpaper',
+    # 'papersize': 'letterpaper',
     # The font size ('10pt', '11pt' or '12pt').
-    #'pointsize': '10pt',
+    # 'pointsize': '10pt',
     # Additional stuff for the LaTeX preamble.
-    #'preamble': '',
+    # 'preamble': '',
 }
 
 # Grouping the document tree into LaTeX files. List of tuples
@@ -209,8 +208,8 @@ latex_documents = [
     (
         "index",
         "pip_chill.tex",
-        u"PIP Chill Documentation",
-        u"Ricardo Bánffy",
+        "PIP Chill Documentation",
+        "Ricardo Bánffy",
         "manual",
     ),
 ]
@@ -240,9 +239,7 @@ latex_documents = [
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [
-    ("index", "pip_chill", u"PIP Chill Documentation", [u"Ricardo Bánffy"], 1)
-]
+man_pages = [("index", "pip_chill", "PIP Chill Documentation", ["Ricardo Bánffy"], 1)]
 
 # If true, show URL addresses after external links.
 # man_show_urls = False
@@ -257,8 +254,8 @@ texinfo_documents = [
     (
         "index",
         "pip_chill",
-        u"PIP Chill Documentation",
-        u"Ricardo Bánffy",
+        "PIP Chill Documentation",
+        "Ricardo Bánffy",
         "pip_chill",
         "One line description of project.",
         "Miscellaneous",