+++ pip_chill/__init__.py
@@ -1,10 +1,8 @@
-# -*- coding: utf-8 -*-
-
 from .pip_chill import chill
 
 __author__ = "Ricardo Bánffy"
 __email__ = "rbanffy@gmail.com"
-__version__ = "1.0.1"
+__version__ = "1.0.4"
 
 
-__all__ = [chill.__name__]
+__all__ = [chill.__name__]  # type: ignore[reportUnsupportedDunderAll]