+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+1.0.4 (...)
+------------------
+* Remove dependency to obsolete `pkg_resources`.
+* Dropped support to 3.7, 3.8 and 3.9 as a consequence of relying on `importlib.metadata` and `packaging`
+* Adding support to 3.11, 3.12, 3.13 and 3.14.
+
 1.0.3 (...)
 ------------------
 * Add dependency on `setuptools` for python >= 3.12 to get access to `pkg_resources`.